Add functions to change trends of one part
The original functions for changing e.g. data types of trends only did
so for a whole trend store at a time. Because changing data types of
columns in a table can be a very time consuming activity, you want to do
that one table at a time. This way, the locks on the tables can be
shorter, because you don't have to wait until the tables of all trend
store parts are changed.