Fix bug in trend_directory.modified updating
Overlapping data loading transactions could cause changes to the data to be
missed by materializations because the trend_directory.modified table would not
be updated if the last finished transaction contained an earlier modified
timestamp. This is fixed by also allowing the `last` field to be updated to an
older timestamp.